+++ src/main/kotlin/no/nav/familie/ef/sak/vedtak/VedtakController.kt
@@ -49,7 +49,8 @@ class VedtakController(
     private val vedtakHistorikkService: VedtakHistorikkService,
     private val behandlingRepository: BehandlingRepository,
     private val nullstillVedtakService: NullstillVedtakService,
-    private val featureToggleService: FeatureToggleService
+    private val featureToggleService: FeatureToggleService,
+    private val angreSendTilBeslutterService: AngreSendTilBeslutterService
 ) {
 
     private val logger = LoggerFactory.getLogger(javaClass)
@@ -70,10 +71,9 @@ class VedtakController(
     @PostMapping("/{behandlingId}/angre-send-til-beslutter")
     fun angreSendTilBeslutter(@PathVariable behandlingId: UUID): Ressurs<UUID> {
         feilHvisIkke(featureToggleService.isEnabled(Toggle.ANGRE_SEND_TIL_BESLUTTER), HttpStatus.SERVICE_UNAVAILABLE) { "Feature toggle for angre send er ikke påskrudd." }
-        val saksbehandling = behandlingService.hentSaksbehandling(behandlingId)
-        tilgangService.validerTilgangTilBehandling(saksbehandling, AuditLoggerEvent.UPDATE)
-        vedtakService.angreSendTilBeslutter(saksbehandling)
-        return Ressurs.success(saksbehandling.id)
+        tilgangService.validerTilgangTilBehandling(behandlingId, AuditLoggerEvent.UPDATE)
+        angreSendTilBeslutterService.angreSendTilBeslutter(behandlingId)
+        return Ressurs.success(behandlingId)
     }
 
     @PostMapping("/{behandlingId}/beslutte-vedtak")
