+++ src/main/kotlin/no/nav/familie/ef/sak/vedtak/VedtakService.kt
@@ -1,16 +1,5 @@
 package no.nav.familie.ef.sak.vedtak
 
-import no.nav.familie.ef.sak.behandling.BehandlingService
-import no.nav.familie.ef.sak.behandling.Saksbehandling
-import no.nav.familie.ef.sak.behandling.domain.BehandlingStatus
-import no.nav.familie.ef.sak.behandlingsflyt.steg.StegType
-import no.nav.familie.ef.sak.behandlingsflyt.task.FerdigstillOppgaveTask
-import no.nav.familie.ef.sak.behandlingsflyt.task.OpprettOppgaveTask
-import no.nav.familie.ef.sak.behandlingshistorikk.BehandlingshistorikkService
-import no.nav.familie.ef.sak.behandlingshistorikk.domain.StegUtfall
-import no.nav.familie.ef.sak.infrastruktur.exception.feilHvis
-import no.nav.familie.ef.sak.infrastruktur.sikkerhet.SikkerhetContext
-import no.nav.familie.ef.sak.oppgave.OppgaveService
 import no.nav.familie.ef.sak.repository.findAllByIdOrThrow
 import no.nav.familie.ef.sak.repository.findByIdOrThrow
 import no.nav.familie.ef.sak.tilkjentytelse.TilkjentYtelseRepository
@@ -20,12 +9,8 @@ import no.nav.familie.ef.sak.vedtak.dto.VedtakDto
 import no.nav.familie.ef.sak.vedtak.dto.tilVedtak
 import no.nav.familie.ef.sak.vedtak.dto.tilVedtakDto
 import no.nav.familie.kontrakter.felles.ef.StønadType
-import no.nav.familie.kontrakter.felles.oppgave.Oppgavetype
-import no.nav.familie.prosessering.internal.TaskService
 import org.springframework.data.repository.findByIdOrNull
-import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDate
 import java.time.YearMonth
 import java.util.UUID
@@ -34,10 +19,6 @@ import java.util.UUID
 class VedtakService(
     private val vedtakRepository: VedtakRepository,
     private val tilkjentYtelseRepository: TilkjentYtelseRepository,
-    private val oppgaveService: OppgaveService,
-    private val behandlingService: BehandlingService,
-    private val behandlingshistorikkService: BehandlingshistorikkService,
-    private val taskService: TaskService
 ) {
 
     fun lagreVedtak(vedtakDto: VedtakDto, behandlingId: UUID, stønadstype: StønadType): UUID {
@@ -118,62 +99,6 @@ class VedtakService(
 
     fun hentHarAktivtVedtak(behandlingId: UUID, localDate: LocalDate = LocalDate.now()): Boolean {
         return hentVedtak(behandlingId).erVedtakAktivtForDato(localDate)
-    }
-
-    @Transactional
-    fun angreSendTilBeslutter(saksbehandling: Saksbehandling) {
-        val vedtak = hentVedtak(behandlingId = saksbehandling.id)
-
-        validerKanAngreSendTilBeslutter(saksbehandling, vedtak)
-
-        behandlingService.oppdaterStegPåBehandling(saksbehandling.id, steg = StegType.SEND_TIL_BESLUTTER)
-        behandlingService.oppdaterStatusPåBehandling(saksbehandling.id, status = BehandlingStatus.UTREDES)
-        behandlingshistorikkService.opprettHistorikkInnslag(
-            behandlingId = saksbehandling.id,
-            stegtype = saksbehandling.steg,
-            utfall = StegUtfall.ANGRE_SEND_TIL_BESLUTTER,
-            metadata = null
-        )
-
-        ferdigstillGodkjenneVedtakOppgave(saksbehandling)
-        opprettBehandleSakOppgave(saksbehandling)
-    }
-
-    private fun opprettBehandleSakOppgave(saksbehandling: Saksbehandling) {
-        taskService.save(
-            OpprettOppgaveTask.opprettTask(
-                OpprettOppgaveTask.OpprettOppgaveTaskData(
-                    behandlingId = saksbehandling.id,
-                    oppgavetype = Oppgavetype.BehandleSak,
-                    beskrivelse = "Angret send til beslutter"
-                )
-            )
-        )
-    }
-
-    private fun ferdigstillGodkjenneVedtakOppgave(saksbehandling: Saksbehandling) {
-        oppgaveService.hentOppgaveSomIkkeErFerdigstilt(oppgavetype = Oppgavetype.GodkjenneVedtak, saksbehandling)?.let {
-            taskService.save(
-                FerdigstillOppgaveTask.opprettTask(
-                    behandlingId = saksbehandling.id,
-                    oppgavetype = Oppgavetype.GodkjenneVedtak,
-                    it.gsakOppgaveId,
-                    personIdent = null
-                )
-            )
-        }
-    }
-
-    private fun validerKanAngreSendTilBeslutter(saksbehandling: Saksbehandling, vedtak: Vedtak) {
-        val innloggetSaksbehandler = SikkerhetContext.hentSaksbehandler()
-        feilHvis(vedtak.saksbehandlerIdent != innloggetSaksbehandler, httpStatus = HttpStatus.BAD_REQUEST) { "Kan ikke angre send til beslutter om du ikke er saksbehandler på vedtaket" }
-        feilHvis(saksbehandling.steg != StegType.BESLUTTE_VEDTAK, httpStatus = HttpStatus.BAD_REQUEST) { "Kan ikke angre send til beslutter når behandling er i steg ${saksbehandling.steg}" }
-        feilHvis(saksbehandling.status != BehandlingStatus.FATTER_VEDTAK, httpStatus = HttpStatus.BAD_REQUEST) { "Kan ikke angre send til beslutter når behandlingen har status ${saksbehandling.status}" }
-
-        val efOppgave = oppgaveService.hentOppgaveSomIkkeErFerdigstilt(oppgavetype = Oppgavetype.GodkjenneVedtak, saksbehandling = saksbehandling) ?: error("Fant ingen godkjenne vedtak oppgave")
-        val tilordnetRessurs = oppgaveService.hentOppgave(efOppgave.gsakOppgaveId).tilordnetRessurs
-        val oppgaveErTilordnetEnAnnenSaksbehandler = tilordnetRessurs != null && tilordnetRessurs != innloggetSaksbehandler
-        feilHvis(oppgaveErTilordnetEnAnnenSaksbehandler, httpStatus = HttpStatus.BAD_REQUEST) { "Kan ikke angre send til beslutter når oppgave er plukket av $tilordnetRessurs" }
     }
 }
